Add WPE favicon support
This adds favicon support to the WPE backend, based on the existing code
introduced by GTK.

A new `WebKitFavicon` type is introduced as the return value for WPE's
favicon calls, which exposes the favicon size and its content through
separate functions.

The favicon data itself is unmodified and in BGRA format.