Add WPE context menu support
This replaces the existing context menu stub implementation with a
minimal version which emits the WebView's context menu signal.

This also resolves a bug where all mouse input would break after
right-clicking, since the stub implementation was waiting for the
context menu to be opened before processing new mouse events.

The `context-menu-dismissed` signal is removed from the WPE backend in a
breaking change, since dismissal is handled by the consumer of the API.