Improve text search colors
This is a hacky solution to fix the suboptimal default search match
colors.

The general matches are identified as such and were already using a
constant with some additional indirection, so this constant was just
updated and the indirection removed.

The primary match is just marked as selection, which means that with the
current implementation it is not possible to give it a different color
than the text selection. Since selection does not work on touch devices
yet anyway, this has been conveniently ignored for now.